feat: migrate code snippet component
