+++ src/components/CodeSnippet/index.ts
@@ -1,0 +1,1 @@
+export { CodeSnippet } from './CodeSnippet'
